add tests for visible_unit, remove unused cansee_ex
