+++ src/console.c
@@ -246,15 +246,17 @@ int lua_do(lua_State * L)
 {
     int status = loadline(L);
     if (status != -1) {
-        if (status == 0)
+        if (status == 0) {
             status = docall(L, 0, 0);
+        }
         report(L, status);
         if (status == 0 && lua_gettop(L) > 0) {     /* any result to print? */
             lua_getglobal(L, "print");
             lua_insert(L, 1);
-            if (lua_pcall(L, lua_gettop(L) - 1, 0, 0) != 0)
+            if (lua_pcall(L, lua_gettop(L) - 1, 0, 0) != 0) {
                 l_message(NULL, lua_pushfstring(L, "error calling `print' (%s)",
-                lua_tostring(L, -1)));
+                    lua_tostring(L, -1)));
+            }
         }
     }
     lua_settop(L, 0);             /* clear stack */
