+++ src/laws.test.c
@@ -1679,25 +1679,6 @@ static void test_cansee(CuTest *tc) {
     test_cleanup();
 }
 
-static void test_cansee_spell(CuTest *tc) {
-    unit *u2;
-    faction *f;
-
-    test_setup();
-    f = test_create_faction(0);
-    u2 = test_create_unit(test_create_faction(0), test_create_region(0, 0, 0));
-
-    CuAssertTrue(tc, cansee_ex(f, u2->region, u2, 0, seen_spell));
-    CuAssertTrue(tc, cansee_ex(f, u2->region, u2, 0, seen_battle));
-
-    set_level(u2, SK_STEALTH, 1);
-    CuAssertTrue(tc, !cansee_ex(f, u2->region, u2, 0, seen_spell));
-    CuAssertTrue(tc, cansee_ex(f, u2->region, u2, 1, seen_spell));
-    CuAssertTrue(tc, cansee_ex(f, u2->region, u2, 1, seen_battle));
-
-    test_cleanup();
-}
-
 static void test_cansee_ring(CuTest *tc) {
     unit *u, *u2;
     item_type *itype[2];
@@ -1835,7 +1816,6 @@ CuSuite *get_laws_suite(void)
     SUITE_ADD_TEST(suite, test_cansee);
     SUITE_ADD_TEST(suite, test_cansee_ring);
     SUITE_ADD_TEST(suite, test_cansee_sphere);
-    SUITE_ADD_TEST(suite, test_cansee_spell);
 
     return suite;
 }
